pyproject.toml
[project]
name = "boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "huggingface-hub>=1.5.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pyyaml>=6.0.3",
    "python-Levenshtein>=0.25.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.0",
]
